Moved raw data folder to backend.
