+++ package.json
@@ -14,6 +14,7 @@
   "dependencies": {
     "@pinecone-database/pinecone": "^5.1.2",
     "@prisma/client": "^6.7.0",
+    "@xenova/transformers": "^2.17.2",
     "cors": "^2.8.5",
     "dotenv": "^16.5.0",
     "express": "^5.1.0",
